get scene method in ucmapeditor.xaml.cs

+++ JDR.Modele/Scene.cs
@@ -13,7 +13,7 @@ namespace JDR.Model
         public Illustration? Background { get; set; }
         public IList<NPC> Characters { get; set; }
         public IList<Obstacle> Obstacles { get; set; }
-        public decimal ZoomValue { get; set; }
+        public double ZoomValue { get; set; }
 
         public int XMapTranslation { get; set; }
 		public int YMapTranslation { get; set; }
